feat: export grid as module
