+++ packages/workspace-fusion/package.json
@@ -43,6 +43,9 @@
 			],
 			"data-source": [
 				"dist/types/lib/integrations/data-source/index"
+			],
+			"grid-module": [
+				"dist/types/modules/grid/index"
 			]
 		}
 	},
@@ -54,7 +57,8 @@
 		"./power-bi": "./dist/esm/lib/integrations/power-bi/index.js",
 		"./grid": "./dist/esm/lib/integrations/grid/index.js",
 		"./data-source": "./dist/esm/lib/integrations/data-source/index.js",
-		"./sidesheet": "./dist/esm/lib/integrations/sidesheet/index.js"
+		"./sidesheet": "./dist/esm/lib/integrations/sidesheet/index.js",
+		"./grid-module": "./dist/esm/modules/grid/index.js"
 	},
 	"dependencies": {
 		"@equinor/eds-core-react": "0.27.0",
