--- packages/workspace-fusion/src/lib/integrations/grid/components/index.ts
@@ -1,3 +1,0 @@
-export * from './addGrid';
-export * from './workspaceHeader';
-export * from './wrapper';
