--- packages/workspace-fusion/src/lib/integrations/grid/components/workspaceHeader/index.ts
@@ -1,1 +1,0 @@
-export { GridHeader } from './GridWorkspaceHeader';
