--- packages/workspace-fusion/src/lib/integrations/grid/components/wrapper/index.ts
@@ -1,1 +1,0 @@
-export * from './GridWrapper';
