--- packages/workspace-fusion/src/lib/integrations/grid/icons/GridIcon.tsx
@@ -1,14 +1,0 @@
-export const GridIcon = (): JSX.Element => {
-	return (
-		<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-			<rect width="24" height="24" />
-
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M2 5C2 4.44772 2.44772 4 3 4H21C21.5523 4 22 4.44772 22 5V19L22 19.0051L22 19.0101C22 19.5624 21.5523 20.0101 21 20.0101H3C2.44772 20.0101 2 19.5624 2 19.0101L2.00001 19.0051L2 19V5ZM20 18.0101V15.3401H17.5V18.0101H20ZM20 13.3401V10.67H17.5L17.5 13.3401H20ZM20 8.67004V6H17.5V8.67004H20ZM13 6H15.5V8.67004H13V6ZM13 10.67H15.5L15.5 13.3401H13L13 10.67ZM13 15.3401H15.5V18.0101H13V15.3401ZM11 18.0101V15.3401H8.5V18.0101H11ZM11 13.3401L11 10.67H8.5L8.5 13.3401H11ZM11 8.67004V6H8.5V8.67004H11ZM4 6H6.5V8.67004H4L4 6ZM4 10.67H6.5L6.5 13.3401H4L4 10.67ZM4 15.3401H6.5V18.0101H4L4 15.3401Z"
-				fill="#6F6F6F"
-			/>
-		</svg>
-	);
-};
