+++ packages/workspace-fusion/src/lib/integrations/grid/index.ts
@@ -1,4 +1,2 @@
 export * from './grid';
 export * from './types';
-export * from './components';
-export * from './utils';
