--- packages/workspace-fusion/src/lib/integrations/grid/utils/index.ts
@@ -1,5 +1,0 @@
-export * from './configureBookmark';
-export * from './configureDataChange';
-export * from './configureHighlightSelection';
-export * from './defaultColDefs';
-export * from './setConfigOnController';
