+++ packages/workspace-fusion/src/lib/types/workspaceProps.ts
@@ -16,7 +16,7 @@ type WorkspaceBaseProps<
 	workspaceOptions: WorkspaceConfig<TData>;
 	customTabs?: CustomTab[];
 	contextOptions?: (filteredData: TData[]) => TContext;
-	modules?: FusionWorkspaceModule<TData>[];
+	modules?: any[];
 	//TODO: Modules?
 };
