+++ packages/workspace-fusion/src/modules/grid/utils/configureDataChange.ts
@@ -1,6 +1,6 @@
 import { GridController } from '@equinor/workspace-ag-grid';
 import { BaseEvent } from '@equinor/workspace-core';
-import { FusionMediator } from '../../../types';
+import { FusionMediator } from '../../../lib/types';
 
 /**Update data on gridController whenever filtered data on mediator changes */
 export function dataChangeEffect<
