+++ packages/workspace-fusion/src/modules/grid/utils/snapShotGridState.ts
@@ -1,5 +1,5 @@
 import { GridController } from '@equinor/workspace-ag-grid';
-import { GridBookmark } from '../../../types';
+import { GridBookmark } from '../../../lib';
 
 export function snapshotGridState<TData extends Record<PropertyKey, unknown>>(
 	gridController: GridController<TData>
